pipeline refactoring: keypath and vars
