+++ core/services/pipeline/task.http_test.go
@@ -166,7 +166,9 @@ func TestHTTPTask_Variables(t *testing.T) {
 			}
 			task.HelperSetDependencies(cfg, db, uuid.UUID{})
 
-			test.vars.Set("meta", test.meta)
+			err = test.vars.Set("meta", test.meta)
+			require.NoError(t, err)
+
 			result, runInfo := task.Run(context.Background(), logger.TestLogger(t), test.vars, test.inputs)
 			assert.False(t, runInfo.IsPending)
 			assert.False(t, runInfo.IsRetryable)
