+++ core/services/pipeline/variables.go
@@ -1,7 +1,6 @@
 package pipeline
 
 import (
-	"bytes"
 	"regexp"
 	"strconv"
 	"strings"
@@ -11,8 +10,8 @@ import (
 
 var (
 	ErrKeypathNotFound = errors.New("keypath not found")
-	ErrKeypathTooDeep  = errors.New("keypath too deep (maximum 2 keys)")
 	ErrVarsRoot        = errors.New("cannot get/set the root of a pipeline.Vars")
+	ErrVarsSetNested   = errors.New("cannot set a nested key of a pipeline.Vars")
 
 	variableRegexp = regexp.MustCompile(`\$\(\s*([a-zA-Z0-9_\.]+)\s*\)`)
 )
@@ -21,6 +20,8 @@ type Vars struct {
 	vars map[string]interface{}
 }
 
+// NewVarsFrom creates new Vars from the given map.
+// If the map is nil, a new map instance will be created.
 func NewVarsFrom(m map[string]interface{}) Vars {
 	if m == nil {
 		m = make(map[string]interface{})
@@ -28,115 +29,74 @@ func NewVarsFrom(m map[string]interface{}) Vars {
 	return Vars{vars: m}
 }
 
-func (vars Vars) Copy() Vars {
-	m := make(map[string]interface{})
-	for k, v := range vars.vars {
-		m[k] = v
-	}
-	return Vars{vars: m}
-}
-
+// Get returns the value for the given keypath or error.
+// The keypath can consist of one or two parts, e.g. "foo" or "foo.bar".
+// The second part of the keypath can be an index of a slice.
 func (vars Vars) Get(keypathStr string) (interface{}, error) {
-	keypath, err := newKeypathFromString(keypathStr)
+	keypathStr = strings.TrimSpace(keypathStr)
+	keypath, err := NewKeypathFromString(keypathStr)
 	if err != nil {
 		return nil, err
 	}
-
-	numParts := keypath.NumParts()
-
-	if numParts == 0 {
+	if keypath.NumParts == 0 {
 		return nil, ErrVarsRoot
 	}
 
 	var val interface{}
 	var exists bool
 
-	if numParts >= 1 {
-		val, exists = vars.vars[string(keypath[0])]
+	if keypath.NumParts >= 1 {
+		val, exists = vars.vars[keypath.Part0]
 		if !exists {
-			return nil, errors.Wrapf(ErrKeypathNotFound, "key %v / keypath %v", string(keypath[0]), keypath.String())
+			return nil, errors.Wrapf(ErrKeypathNotFound, "key %v / keypath %v", keypath.Part0, keypathStr)
 		}
 	}
 
-	if numParts == 2 {
+	if keypath.NumParts == 2 {
 		switch v := val.(type) {
 		case map[string]interface{}:
-			val, exists = v[string(keypath[1])]
+			val, exists = v[keypath.Part1]
 			if !exists {
-				return nil, errors.Wrapf(ErrKeypathNotFound, "key %v / keypath %v", string(keypath[1]), keypath.String())
+				return nil, errors.Wrapf(ErrKeypathNotFound, "key %v / keypath %v", keypath.Part1, keypathStr)
 			}
 		case []interface{}:
-			idx, err := strconv.ParseInt(string(keypath[1]), 10, 64)
+			idx, err := strconv.ParseInt(keypath.Part1, 10, 64)
 			if err != nil {
 				return nil, errors.Wrapf(ErrKeypathNotFound, "could not parse key as integer: %v", err)
-			} else if idx > int64(len(v)-1) {
-				return nil, errors.Wrapf(ErrKeypathNotFound, "index %v out of range (length %v / keypath %v)", idx, len(v), keypath.String())
+			} else if idx < 0 || idx > int64(len(v)-1) {
+				return nil, errors.Wrapf(ErrKeypathNotFound, "index %v out of range (length %v / keypath %v)", idx, len(v), keypathStr)
 			}
 			val = v[idx]
 		default:
-			return nil, errors.Wrapf(ErrKeypathNotFound, "value at key '%v' is a %T, not a map or slice", string(keypath[0]), val)
+			return nil, errors.Wrapf(ErrKeypathNotFound, "value at key '%v' is a %T, not a map or slice", keypath.Part0, val)
 		}
 	}
 
 	return val, nil
 }
 
-func (vars Vars) Set(dotID string, value interface{}) {
+// Set sets a top-level variable specified by dotID.
+// Returns error if either dotID is empty or it is a compound keypath.
+func (vars Vars) Set(dotID string, value interface{}) error {
 	dotID = strings.TrimSpace(dotID)
 	if len(dotID) == 0 {
-		panic(ErrVarsRoot)
-	} else if strings.IndexByte(dotID, keypathSeparator[0]) >= 0 {
-		panic("cannot set a nested key of a pipeline.Vars")
+		return ErrVarsRoot
+	} else if strings.Contains(dotID, string(KeypathSeparator[0])) {
+		return errors.Wrapf(ErrVarsSetNested, "%s", dotID)
 	}
+
 	vars.vars[dotID] = value
+
+	return nil
 }
 
-type Keypath [2][]byte
-
-var keypathSeparator = []byte(".")
-
-func newKeypathFromString(keypathStr string) (Keypath, error) {
-	if len(keypathStr) == 0 {
-		return Keypath{}, nil
+// Copy makes a copy of Vars by copying the underlying map.
+// Used by scheduler for new tasks to avoid data races.
+func (vars Vars) Copy() Vars {
+	newVars := make(map[string]interface{})
+	// No need to copy recursively, because only the top-level map is mutable (see Set()).
+	for k, v := range vars.vars {
+		newVars[k] = v
 	}
-	// The bytes package uses platform-dependent hardware optimizations and
-	// avoids the extra allocations that are required to work with strings.
-	// Keypaths have to be parsed quite a bit, so let's do it well.
-	kp := []byte(keypathStr)
-
-	n := 1 + bytes.Count(kp, keypathSeparator)
-	if n > 2 {
-		return Keypath{}, errors.Wrapf(ErrKeypathTooDeep, "while parsing keypath '%v'", keypathStr)
-	}
-	idx := bytes.IndexByte(kp, keypathSeparator[0])
-	if idx == -1 || idx == len(kp)-1 {
-		return Keypath{kp, nil}, nil
-	}
-	return Keypath{kp[:idx], kp[idx+1:]}, nil
-}
-
-func (keypath Keypath) NumParts() int {
-	switch {
-	case keypath[0] == nil && keypath[1] == nil:
-		return 0
-	case keypath[0] != nil && keypath[1] == nil:
-		return 1
-	case keypath[0] == nil && keypath[1] != nil:
-		panic("invariant violation: keypath part 1 is non-nil but part 0 is nil")
-	default:
-		return 2
-	}
-}
-
-func (keypath Keypath) String() string {
-	switch keypath.NumParts() {
-	case 0:
-		return "(empty)"
-	case 1:
-		return string(keypath[0])
-	case 2:
-		return string(keypath[0]) + string(keypathSeparator) + string(keypath[1])
-	default:
-		panic("invariant violation: keypath must have 0, 1, or 2 parts")
-	}
+	return NewVarsFrom(newVars)
 }
